+++ CopilotKit/CopilotKit.code-workspace
@@ -1,24 +1,24 @@
 {
-	"folders": [
-		{
-			"path": "."
-		}
-	],
-	"settings": {
-		"jest.disabledWorkspaceFolders": [
-			"CopilotKit",
-			"eslint-config-custom",
-			"cloud",
-			"react-textarea",
-			"shared",
-			"tsconfig",
-			"react-ui",
-			"react-core",
-			"tailwind-config",
-			"next-openai",
-			".next",
-			"CopilotKit",
-			"types"
-		]
-	}
+  "folders": [
+    {
+      "path": "."
+    }
+  ],
+  "settings": {
+    "jest.disabledWorkspaceFolders": [
+      "CopilotKit",
+      "eslint-config-custom",
+      "backend",
+      "react-textarea",
+      "shared",
+      "tsconfig",
+      "react-ui",
+      "react-core",
+      "tailwind-config",
+      "next-openai",
+      ".next",
+      "CopilotKit",
+      "types"
+    ]
+  }
 }
