+++ CopilotKit/examples/next-openai/package.json
@@ -10,7 +10,7 @@
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist && rm -rf .next"
   },
   "dependencies": {
-    "@copilotkit/cloud": "workspace:*",
+    "@copilotkit/backend": "workspace:*",
     "@copilotkit/shared": "workspace:*",
     "@copilotkit/react-core": "workspace:*",
     "@copilotkit/react-textarea": "workspace:*",
