+++ CopilotKit/examples/next-openai/src/app/api/copilotkit/openai/route.ts
@@ -1,4 +1,4 @@
-import { CopilotKitBackend } from "@copilotkit/cloud";
+import { CopilotKitBackend } from "@copilotkit/backend";
 
 export const runtime = "edge";
