+++ CopilotKit/packages/backend/src/lib/openai-adapter.ts
