+++ CopilotKit/packages/backend/src/streams/assistant-response.ts
