+++ CopilotKit/packages/backend/src/types/service-adapter.ts
