+++ CopilotKit/packages/backend/src/utils/openai-stream-interceptor.ts
