+++ CopilotKit/packages/backend/src/utils/openai.ts
