+++ CopilotKit/packages/backend/src/utils/utils.ts
